Extract RecipeImageCell
